requirements.txt
mcp[cli]>=0.1.0
Pillow>=10.0.0
pywinauto>=0.6.8
asyncio>=3.4.3
pywin32>=306
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.1
jinja2  
websockets>=12.0
